// libs/enums/clothes.enum.ts
export enum ClotheCategory {
  T_SHIRT = "T_SHIRT",
  HOODIE = "HOODIE",
  SHORTS = "SHORTS",
  PANTS = "PANTS",
  SHOES = "SHOES",
  GLOVES = "GLOVES",
  SOCKS = "SOCKS",
  JACKET = "JACKET",
  SPORTS_BRA = "SPORTS_BRA",
}

export enum ClotheStatus {
  ACTIVE = "ACTIVE",
  SOLD = "SOLD",
  DELETE = "DELETE",
}

export enum ClotheMaterial {
  COTTON = "COTTON",
  POLYESTER = "POLYESTER",
  NYLON = "NYLON",
  SPANDEX = "SPANDEX",
  FLEECE = "FLEECE",
  MESH = "MESH",
}


export enum ClotheSize {
  XS = "XS",
  S = "S",
  M = "M",
  L = "L",
  XL = "XL",
  XXL = "XXL",
}

export enum ClotheGender {
  MEN = "MEN",
  WOMEN = "WOMEN",
  UNISEX = "UNISEX",
}

export enum ClotheColor {
  BLACK = "BLACK",
  WHITE = "WHITE",
  GRAY = "GRAY",
  RED = "RED",
  BLUE = "BLUE",
  GREEN = "GREEN",
  YELLOW = "YELLOW",
  NAVY = "NAVY",
  PINK = "PINK",
}